pyproject.toml
[tool.poetry]
name = "repl-nix-jovian-careers-website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
